For purpose of QA extract xOffset=0 FiberTrace from sum, rec, snr, norm
Create new FiberTraceSets containing the extracted 0-offset FiberTrace
from the summed-up image, the reconstructed image, the S/N image, and
the normalized Flat. If there are systematics, they are most likely in
individual columns of the (in the original images curved) FiberTraces.

Enable display of the 0-offset FiberTraceSets for visual checking.