Add defects to existing output file
We used to completely rewrite the output file but then the output file
would only contain the defects of the certain CCD we were looking
at. Now we read in the output file if it already exists and only replace
the defect list for the certain CCD while keeping the defects of the
other CCDs.