constructFiberFlatTask.py: Add xOffsetHdrKeyWord
The simulations used the fits header keyword 'sim.slit.xoffset'.
The files from LAM have 'W_FCA_SHIFT'. For now we just
add a config parameter which can be used to change the
header keyword the task is looking for. However, until all
fits files and test data are updated, we leave the default
at 'sim.slit.xoffset' because otherwise we would have to
update the unit tests, quick-start guide, and runPipeline.