Change doc string for xOffsetHdrKeyWord
The LAM hexapod offset keywords were called SHIFT, DITHER, FOCUS.
The reason is that X,Y,and Z are ill-defined. This is worse in PFS,
since the detector is physically on its side, with the spectra
running left-right. We therefore change the doc string for the
config parameter "xOffsetHdrKeyWord" from "xoffset" to "fiberOffset".