constructFiberTrace: Move trace and profile determination to subtasks
So far tracing of the FiberTrace apertures as well as the profile
determination was done by declaring the corresponding tasks, setting
their config parameters, and executing the tasks manually. This had
the disadvantage that the config parameters of those tasks were
included in ConstructFiberTraceConfig. The better solution is to
run those tasks as subtasks and read their respective config files
instead of repeating the config parameter declarations. This way
we can simply read the standard config files which can be shared
between different tasks.