genDefectFits.py: Use CCD Name as HDU Header Name
So far PfsMapper._extractDetectorName returned _extractDetectorId
instead of the actual name to get defects working, as
cameraMapper.bypass_defects looks for the CCD Name in the HDU
Headers whild genDefectFits.py used the CCD Id for the HDU Header
Name. We now fix this in genDefectFits.py and use the CCD Name
for the HDU Header Name.