First crude cut at working cosmic ray conditions.
This is slightly tricky, as the PSF is undersampled by ~2. Even after
the coeffs are set to not match arc lines, CRs tracks crossing fiber
traces are prone to "leak" along the traces.

This is less tricky than it appears with simulated data, which
currently has no scattering. The committed config allows simulated sky
spectra to clear without too many false positives.