PfsMapper: update I/O of wrapped types
SpectrumSet is written as a PfsArm.
FiberTraceSet is written as a PfsFiberTrace.
DetectorMap now has the I/O code included within the class rather than
using a separate class.
PfsConfig is the outlier since it has a separate wrapper class providing
the I/O layer; that's only because we don't have a drp_stella C++ version
of it.