PfsMapper: use visit0 in flat,fibertrace
It's already used in detectorMap, but we now have the ability to use it
in flat and fibertrace too. Including it in bias and dark is possible,
but it requires overriding the top-level task which means using a new
executable instead of constructBias.py/constructDark.py. We may want
to get the capability in upstream instead.