add AMBRE spectra
The pipeline needs to know the spectra of calibration sources
for the purpose of flux calibration. We're currently cheating
by knowing which spectrum was inserted in the sims (because we
don't currently have an algorithm for selecting a suitable
spectrum based on the extracted spectrum), but may in the
future use these spectra more generally as well.