ingest: deal with undefined values
LSST 18.1.0 now handles undefined values in the FITS headers,
but the ingest script treats them as None and blows up.
Handle these cases explicitly.

This involves copying code from LSST in order to fix it: not
desirable, but unavoidable. Not a huge deal in this instance,
as this code will probably disappear with the next LSST
upgrade as we move to the Gen3 middleware.