config: modify cosmicray cond3 values
Condition 3 in the cosmic ray identification code is the test for
sharpness.

In simulated images with full fiber density, many of the traces
are being identified as CRs. This appears to be because cond3_fac
is set too low. The limit on sharpness is set using a PSF realised
at the center of a pixel, but PSFs that are not centered on the
pixel can have even sharper edges. Increasing cond3_fac allows for
this extra sharpness.

The condition 3 parameter values used for HSC (which are the defaults)
do a much better job of not being overwhelmed by the sharp edges of the
fiber traces.