support intra-day ranges for calib validity
This involves modifying the calib registry and ingest process, and
reconfiguring the calib construction.

Multiple functions in the LSST ingest code need to be modified away from
assuming whole-day ranges.

Following this change, dateObs (in the image registry) and calibDate (in
the calib registry) will be dates, used for filenames; taiObs (in the image
registry) and calibTime (in the calib registry) will be used for calib
validity ranges.