adjust to fiberProfiles replacing fiberTrace
FiberTrace contains both the profile and the position of the trace, but
the position of the trace is also in the detectorMap. Furthermore, because
the FiberTrace contains the profile realised for a particular trace
position, it's not possible to realise the profile at a different position
(it's undersampled, so we can't just resample). In order to be able to
realise the trace at a different position, we need to store the profile
(FiberProfile), and not the realised profile (FiberTrace).

This results in several related changes across the package:
* calib construction script names have changed
* calib type name has changed (fiberTrace --> fiberProfiles)
* FindAndTraceAperturesTask has been replaced by BuildFiberProfilesTask
* use MeasureSlitOffsetsTask in reduceArc (can't use it in reduceExposure
  yet because we don't have a sky lines list)