ingest: update calibs if already registered
We often want to overwrite a registered calib (e.g., we start with
a detectorMap from the simulator that we want to replace with a
detectorMap we measured ourselves), but we had no way of doing
that without directly modifying the sqlite database, and using
the register.ignore=True configuration option doesn't modify
the rows to point to the new calib we want to use.

Added the ability to update a row in the calib database, which
is used when a new file overlaps with an existing entry.