Port LSST fix of DM-29117
This allows changing the type of a scalar entry in a
PropertyList or PropertySet. We need this because we
fix the pfsDesignId value (W_PFDSGN) in headers, where
they sometimes go from an int to a long.

This is done by copying daf_base's (modern) file
python/lsst/daf/base/propertyContainer/propertyContainerContinued.py,
stripping out a few things that we really don't need,
and then monkey-patching it over the broken classes.