flag Kr line causing problems
This line has 1/1000th of the flux of a neighbouring line
(according to NIST) which is just beyond the 0.3nm exclusion
zone. We're not measuring its position accurately, leading
to awful distortion fits. It's not clipped by a S/N cut
due to the nearby bright line, so we need to flag it.