isr: mask pixels with a value of zero
Pixels with value of zero were intentionally not read out,
and only exist in the data as padding. We mask zero pixels
before overscan subtraction, and then afterwards reset them
to zero. The final result after ISR is that the zero pixels
have a value close to zero (though not exactly zero due to
bias- and dark-subtraction, and any effect of FITS
compression) and are masked, and the zero-valued pixels in
the overscan are not used for overscan subtraction (this
may necessitate a different choice of overscan subtraction
method than the default).