PfsMapper: make reading of ingested guider files work
    butler.get(guider, visit=70901, sequence=0, hdu=1)

now works after ingesting the guider files.

Clobbered the 'rawAG' dataset type because I think my 'guider'
dataset type is a bit more useful: doesn't need 'arm' and
'spectrograph', but 'sequence' and 'hdu' which is more
guider-specific.

The guider headers are fairly sparse at the moment. The EXTNAME
header keyword identifies the camera, but I'm scared to use it
because it's not clear to me how that extends to multiple
exposures (EXTNAME should be unique within a file). So I
calculate the camera number from the HDU number, and check it
against EXTNAME.