add facility to check/fix headers
The LSST Gen3 butler renames files, so we can't guarantee that
the filenames will have a familiar format. But not all files
have the necessary header keywords that we need if we can't
determine values from the filename. So add a utility that
checks the headers, and fixes them to correspond to the values
we'd get from the filename; we can use this before ingesting
into the butler.