Add regenerated sky linelist with transition and reference source information.
Linelist is generated from Osterbrock(1997) and Rousselot(2000) lists.

The Osterbrock data list transitions, but not relative intensities,
while the Rousselot lists relative intensities but no transition information.
These lists are first merged, and then due to the relatively low spectral resolution
of the PFS (dLambda~0.3 nm) that would otherwise affect the line intensity measurements,
neighbouring lines are grouped. These grouped lines are flagged with bitmask
ReferenceLineStatus.COMBINED.

The transition field in the case of those combined lines
contains all the transitions up to the allowed character limit of 128. Missing transitions
are indicated by an ellipsis (...).

Finally, non-OH lines (HI, Na, OI) are added.

The original reference source of those lines are noted in the source column as bitmasks
in accordance to the ReferenceLineSource class.