update instrument definition for Gen3
The previous version was a placeholder constructed as part of updating
the LSST stack to v23. Now we desire to use it in earnest as part of the
LSST Gen3 middleware.

There is one main instrument definition (intended for use with the actual
PFS at Subaru) and two subclasses (intended for the simulator and
development data from LAM; we might need to add JHU later).

The PfsRawFormatter is based on the Gen2 PfsMapper's manipulations of
raw data.

Formatted file with black.