Add files required for Gen3
butler.yaml defines I/O for the various dataset types.

dimensions.yaml defines the butler registry schema.

skymap_rings.py is the configuration for the HSC rings skymap, for use with
science observations.

skymap_discrete.py is a skymap configuration suitable for use with the
integration test (faster to load into the registry than the rings skymap).

gen3.py contains PfsRawIngestTask, which is the task to use for ingesting
raw PFS files into the datastore: it adds PFS-specific information to the
registry.