use PFS-specific exposure formatter
to suppress warnings about filter mis-match.

We've been getting warnings about a mis-match between the filter
in the Exposure and the filter in the dataId. The filter in the
dataId comes from assuming that all components of an exposure
have the same filter (good for most imaging instruments) but
that's not the case for PFS. So subclassed the formatter that's
producing the warning and substituted it in the butler datastore
configuration.

This involved moving a bit of Gen3 furniture around: moved the
formatters into their own file, and the loadCamera function into
its own file because it's used by both the formatters and the
instrument.