ingest: fix calib validity dates so they include the calib date
With the validityUntilSuperseded setting, we set the validity
times to be from the time of one calib to the time of the
next. But that range does not include that calib because
apparently the comparison uses > for the lower range instead
of >=. To fix this, adjust the validity ranges by a microsecond
so that the validity range for the calib includes the time of
the calib.