PfsAssembleCcdTask: multiply each amplifier by the gain
We want the images to be in units of electrons rather than ADU. I tried
simply setting doApplyGains=True in the IsrTask, but there is a prohibition
against having doApplyGains=True and doFlat=True (because the IsrTask uses
gain application not as a method of changing the units but as an
alternative to flat-fielding); so that is not what we want. Instead, we
convert images to electrons when we assemble the image from amplifiers.
This is after saturation detection (which requires units of ADU) and before
any other image operations, which guarantees that all images will be in
units of electrons.

This invalidates existing biases, darks and flats, because the units have
changed. This is unfortunate, but there is nowhere to hook into in
IsrTask that will allow us to multiply by the gain after bias-subtraction
but before flat-fielding.