ingest: fix postgresql ingest
PostgreSQL ingest appears to have been broken on the LSST side
by DM-21869, which hard-coded an 'INSERT OR IGNORE' sql command
that pgsql doesn't support. The 'addVisits' method disappeared
too as part of that. This requires re-implementing the addRow
method on the PFS side, with the visit adding removed and the
unsupported SQL removed.