Only run special n-band code when processing H4RG data
The bug that caused the tests to fail was caused by some H4RG-only
tests in this routine, which this change prevents.  However, this
code should not have been here in the first place; we should move
the variance code to the ISR. It's not clear why it was put here in
the first place, but I suspect that it was a mistake by RHL before
he wrote the H4RG ISR code.  PIPE2D-1103