move ingestPfsDefects to drp_pfs_data as makePfsDefects
We need to use this script for both Gen2 and Gen3 LSST middleware,
so it's not appropriate to hard-code only one. It's easiest if we
just have this script do the translation, and then call the ingest
appropriate to the middleware manually.
Further, Gen3 requires that the defects are ingested from the
drp_pfs_data package, so we'll put everything we need in that
package.