PfsTranslator: fix default_search_path
The default_search_path is being cast to a list, which broke the
string. I suspect in the recent LSST upgrade, it was converted
from a string to a list. Since we still had it as a string, it
was being sliced into characters, which meant that there was no
full string for the path, which meant that the header fixes
weren't being found.