revise merging of sky lines
The current sky linelist seems to be the main reason we're having
trouble adjusting the detectorMap in at least some cases. The main
reason is that it merges the Osterbrock and Rousselot lists in a
non-principled manner. The Osterbrock list is linked to observations,
while the Rousselot list is a theoretical list without line strengths
(the original paper was focussed on NIR observations, and the linelist
in the optical appears to have gotten tagged on for good measure for
use by ESO). Merging the two lists without regard to the line
identifications or strengths means that the wavelength of the merged
lines can be inaccurate. Furthermore, by merging lines
indiscriminately, it becomes difficult to identify lines that are
blended and should be avoided in our fits.

This revised list is constructed solely from the Osterbock list (for
the wavelength range it covers; otherwise we keep the Rousselot list),
converting air wavelengths to vacuum using the Edlen formula. I've
only merged lines when they are within 0.1 nm and I believe there is
reason to expect them to have the same strength (e/f and P1/P2,
etc.). The intensities come from interpolating Jim Gunn's sky model
(used for the simulator) at the positions of the lines.