skylines: mark bright blue lines as PROTECTED
PROTECTED = 0x400 = 1024 is a new status bit that indicates the line should
not be subject to exclusion zone limitations. The five bright lines in the
blue are precious, and we don't want to throw them away just because there
are really faint OH lines nearby.

With that, we no longer need to flag the OH lines around those bright
lines.