isr: add dichroic background subtraction
The flux in the dichroic region needs to correspond between the
two arms, but may not if there is diffuse excess background.
This commit introduces a feature that measures and removes any
diffuse excess background in the dichroic region. It uses the
fact that the flux from the fibers should be zero deep in the
dichroic, and subtracts the median value in that region from
the image (if there are two dichroics, it subtracts a ramp
across the image).