monkey-patch DM-31448 from LSST
DM-31448 (Allow gen2 butler lookups even when fields in the
dataId are not in the registry) contains a feature that we
need for PFS, but will not be released by LSST (it's a Gen2
fix, and they've removed Gen2). This commit imports this
change, along with an additional change from RHL that
generalises the fix for use with postgres.