Move IPC coefficients from a config parameter to a calibration product
N.b. Logically they'd probably be better as a curatedCalibration product,
but they are naturally FITS files (especially as they'll soon be a large
MEF handling spatial variation).  The _procedure_ docs will have to make
it clear that they are to be ingested from pfs_instdata not a private
rerun