isr: record calib path in image metadata
We've been using a symlink as the calib, which is handy for
remembering and staying current, but it doesn't preserve the
knowledge of exactly what calib was used. Here we provide
a function for figuring that out from a Gen2 butler, and
record the value in the image header for posterity.