add new skyLines file constructed by Christian Jespersen
Christian Kragh Jespersen generated a linelist by taking the
OH lines from pgopher and simulating the spectra under a
variety of conditions. This allows getting an accurate wavelength
for blend complexes, and identifying lines that move under
different conditions (because of varying line strengths of the
blend components).

CKJ will call his code RAGNAR, so put that in the filename.
This list is derived from his file sweep_0.1_1_05122023.csv.
Making a new file instead of overwriting skylines.txt so we
can revert with a configuration change if necessary.