Add basic up-the-ramp processing
We want the H4 parts of the isrTask to have access to the full ramp cube, but
also want the isrTask to remain compatible with lsst.ip.isr, especially for b/r/m.
Our fiddle is to have the butler return a PfsRaw object, which provides
normal 2d images for the visible, and access to the ramp reads for the NIR.

Have isrTask accept PfsRaw instead of Exposure.
Hoist H4 processing configs into .h4 section.
Rotate images immediately when reading.