remove pre-definition of detectorMap_bootstrap
and replace it with a pre-definition of detectorMap_calib.

We don't need to pre-define detectorMap_bootstrap any more, because it's
the result of a pipeline (previously, it was the simulator detectorMap to
be used for bootstrapping, rather than the result of the bootstrapping).
However, we do need to pre-define detectorMap_calib because we're going to
produce those by copying detectorMap_candidate (to avoid cycles in the
pipeline).