gen3: use visit instead of exposure
'visit' is so entrenched that we only want to use that, and not
exposure (deliberately not following LSST's example, because
we have more data).

Renamed the 'exposure' dimension as 'visit', and removed all
other mentions of 'visit'. Copied and edited the ISR connections
from ip_isr, so all the inputs/outputs use 'visit' instead of
'exposure', and for good measure I replaced 'detector' by
'arm','spectrograph' too (and then a couple of inputs need
lookup functions).