Have isrTask accept PfsRaw instead of Exposure
We want the H4 parts of the isrTask to have access to the full ramp cube, but also want the isrTask to remain compatible with lsst.ip.isr, especially for b/r/m.

Our fiddle is to have the butler return a PfsRaw object, which provides normal 2d images for the visible, quicklook 2d images for the NIR, and 3d cubes for the NIR. Only a few routines at the run() level need to handle the changes.