add PfsDark and ImageCube
The NIR darks are going to be a bit more complicated: a cube
of images (one for each read) with images stored in separate
HDUs. We need to provide a proxy object to the PfsIsrTask so
that we can have a single type to handle both CCD and NIR
darks (like we do with PfsRaw); this is PfsDark.