Make PfsTargetSpectraFormatter more general
`PfsTargetSpectraFormatter` is useful when they want to implement
`single` subtype as in `butler.get("TYPE.single")`. But its use was
restricted because it expected the class specified by the variable
`PfsTargetSpectraClass` to be contained in specific modules.

This commit lifts this restriction to enable `PfsTargetSpectraFormatter`
to be used more widely.